+++ FocusMac/Sources/FocusMac/Models/ScanSettings.swift
@@ -63,6 +63,7 @@ public struct ScanSettings: Codable, Sendable {
     public var quality: ExportQualityOption = .matchSource
     public var audioTrackIndex: Int = 0
     public var autoRender: Bool = false
+    public var exportClipsFolder: Bool = false
     public var preventSleep: Bool = true
 
     public init() {}